Move sector-related data structures into `sectors` module
