+++ crates/sc-consensus-subspace/src/block_import.rs
@@ -52,8 +52,9 @@ use sp_runtime::traits::{Block as BlockT, Header as HeaderT, NumberFor, One};
 use sp_runtime::Justifications;
 use std::marker::PhantomData;
 use std::sync::Arc;
+use subspace_core_primitives::sectors::SectorId;
 use subspace_core_primitives::segments::{HistorySize, SegmentHeader, SegmentIndex};
-use subspace_core_primitives::{BlockNumber, PublicKey, SectorId, SolutionRange};
+use subspace_core_primitives::{BlockNumber, PublicKey, SolutionRange};
 use subspace_proof_of_space::Table;
 use subspace_verification::{calculate_block_weight, PieceCheckParams, VerifySolutionParams};
 use tracing::warn;
