+++ crates/subspace-farmer-components/src/plotting.rs
@@ -26,9 +26,10 @@ use std::sync::Arc;
 use std::time::Duration;
 use subspace_core_primitives::crypto::kzg::Kzg;
 use subspace_core_primitives::crypto::{blake3_hash, blake3_hash_parallel, Scalar};
-use subspace_core_primitives::pieces::{PieceIndex, PieceOffset, Record, SBucket};
+use subspace_core_primitives::pieces::{PieceIndex, PieceOffset, Record};
+use subspace_core_primitives::sectors::{SBucket, SectorId, SectorIndex};
 use subspace_core_primitives::segments::HistorySize;
-use subspace_core_primitives::{Blake3Hash, PosSeed, PublicKey, SectorId, SectorIndex};
+use subspace_core_primitives::{Blake3Hash, PosSeed, PublicKey};
 use subspace_erasure_coding::ErasureCoding;
 use subspace_proof_of_space::{Table, TableGenerator};
 use thiserror::Error;
