+++ crates/subspace-farmer-components/src/reading.rs
@@ -18,8 +18,8 @@ use std::simd::Simd;
 use std::str::FromStr;
 use std::{fmt, io};
 use subspace_core_primitives::crypto::{blake3_hash, Scalar};
-use subspace_core_primitives::pieces::{Piece, PieceOffset, Record, SBucket};
-use subspace_core_primitives::SectorId;
+use subspace_core_primitives::pieces::{Piece, PieceOffset, Record};
+use subspace_core_primitives::sectors::{SBucket, SectorId};
 use subspace_erasure_coding::ErasureCoding;
 use subspace_proof_of_space::{Table, TableGenerator};
 use thiserror::Error;
