+++ crates/subspace-farmer/src/cluster/farmer.rs
@@ -29,7 +29,8 @@ use std::sync::Arc;
 use std::time::{Duration, Instant};
 use subspace_core_primitives::crypto::blake3_hash_list;
 use subspace_core_primitives::pieces::{Piece, PieceOffset};
-use subspace_core_primitives::{Blake3Hash, SectorIndex};
+use subspace_core_primitives::sectors::SectorIndex;
+use subspace_core_primitives::Blake3Hash;
 use subspace_farmer_components::plotting::PlottedSector;
 use subspace_rpc_primitives::SolutionResponse;
 use tokio::time::MissedTickBehavior;
