+++ crates/subspace-farmer/src/cluster/plotter.rs
@@ -27,7 +27,8 @@ use std::num::NonZeroUsize;
 use std::pin::{pin, Pin};
 use std::sync::Arc;
 use std::time::{Duration, Instant};
-use subspace_core_primitives::{PublicKey, SectorIndex};
+use subspace_core_primitives::sectors::SectorIndex;
+use subspace_core_primitives::PublicKey;
 use subspace_farmer_components::plotting::PlottedSector;
 use subspace_farmer_components::FarmerProtocolInfo;
 use tokio::sync::{OwnedSemaphorePermit, Semaphore};
