+++ crates/subspace-farmer/src/farm.rs
@@ -19,8 +19,8 @@ use std::sync::Arc;
 use std::time::Duration;
 use std::{fmt, io};
 use subspace_core_primitives::pieces::{Piece, PieceIndex, PieceOffset};
+use subspace_core_primitives::sectors::SectorIndex;
 use subspace_core_primitives::segments::SegmentIndex;
-use subspace_core_primitives::SectorIndex;
 use subspace_farmer_components::auditing::AuditingError;
 use subspace_farmer_components::plotting::PlottedSector;
 use subspace_farmer_components::proving::ProvingError;
