+++ crates/subspace-farmer/src/farm/plotted_pieces.rs
@@ -11,7 +11,7 @@ use std::future::Future;
 use std::hash::Hash;
 use std::sync::Arc;
 use subspace_core_primitives::pieces::{Piece, PieceIndex, PieceOffset};
-use subspace_core_primitives::SectorIndex;
+use subspace_core_primitives::sectors::SectorIndex;
 use subspace_farmer_components::plotting::PlottedSector;
 use tracing::{trace, warn};
