+++ crates/subspace-farmer/src/plotter/gpu/cuda.rs
@@ -7,8 +7,8 @@ use rayon::{ThreadPool, ThreadPoolBuildError, ThreadPoolBuilder};
 use std::sync::atomic::{AtomicBool, Ordering};
 use std::sync::Arc;
 use subspace_core_primitives::pieces::{PieceOffset, Record};
+use subspace_core_primitives::sectors::SectorId;
 use subspace_core_primitives::segments::HistorySize;
-use subspace_core_primitives::SectorId;
 use subspace_farmer_components::plotting::RecordsEncoder;
 use subspace_farmer_components::sector::SectorContentsMap;
 use subspace_proof_of_space_gpu::cuda::CudaDevice;
