+++ crates/subspace-farmer/src/plotter/pool.rs
@@ -5,7 +5,8 @@ use async_trait::async_trait;
 use futures::channel::mpsc;
 use std::any::type_name_of_val;
 use std::time::Duration;
-use subspace_core_primitives::{PublicKey, SectorIndex};
+use subspace_core_primitives::sectors::SectorIndex;
+use subspace_core_primitives::PublicKey;
 use subspace_farmer_components::FarmerProtocolInfo;
 use tracing::{error, trace};
