+++ crates/subspace-farmer/src/single_disk_farm.rs
@@ -69,8 +69,9 @@ use std::{fmt, fs, io, mem};
 use subspace_core_primitives::crypto::kzg::Kzg;
 use subspace_core_primitives::crypto::{blake3_hash, Scalar};
 use subspace_core_primitives::pieces::Record;
+use subspace_core_primitives::sectors::SectorIndex;
 use subspace_core_primitives::segments::{HistorySize, SegmentIndex};
-use subspace_core_primitives::{Blake3Hash, PublicKey, SectorIndex};
+use subspace_core_primitives::{Blake3Hash, PublicKey};
 use subspace_erasure_coding::ErasureCoding;
 use subspace_farmer_components::file_ext::FileExt;
 use subspace_farmer_components::reading::ReadSectorRecordChunksMode;
