+++ crates/subspace-farmer/src/single_disk_farm/piece_reader.rs
@@ -10,7 +10,8 @@ use std::collections::HashSet;
 use std::future::Future;
 use std::sync::Arc;
 use subspace_core_primitives::pieces::{Piece, PieceOffset};
-use subspace_core_primitives::{PublicKey, SectorId, SectorIndex};
+use subspace_core_primitives::sectors::{SectorId, SectorIndex};
+use subspace_core_primitives::PublicKey;
 use subspace_erasure_coding::ErasureCoding;
 use subspace_farmer_components::reading::ReadSectorRecordChunksMode;
 use subspace_farmer_components::sector::{sector_size, SectorMetadataChecksummed};
