+++ crates/subspace-farmer/src/single_disk_farm/plot_cache.rs
@@ -15,7 +15,8 @@ use std::sync::{Arc, Weak};
 use std::{io, mem};
 use subspace_core_primitives::crypto::blake3_hash_list;
 use subspace_core_primitives::pieces::{Piece, PieceIndex};
-use subspace_core_primitives::{Blake3Hash, SectorIndex};
+use subspace_core_primitives::sectors::SectorIndex;
+use subspace_core_primitives::Blake3Hash;
 use subspace_farmer_components::file_ext::FileExt;
 use subspace_farmer_components::sector::SectorMetadataChecksummed;
 use subspace_networking::libp2p::kad::RecordKey;
