+++ crates/subspace-farmer/src/single_disk_farm/plot_cache/tests.rs
@@ -12,8 +12,8 @@ use std::assert_matches::assert_matches;
 use std::num::NonZeroU64;
 use std::sync::Arc;
 use subspace_core_primitives::pieces::{Piece, PieceIndex, Record};
+use subspace_core_primitives::sectors::SectorIndex;
 use subspace_core_primitives::segments::HistorySize;
-use subspace_core_primitives::SectorIndex;
 use subspace_farmer_components::file_ext::FileExt;
 use subspace_farmer_components::sector::{SectorMetadata, SectorMetadataChecksummed};
 use subspace_networking::libp2p::kad::RecordKey;
