+++ crates/subspace-farmer/src/single_disk_farm/plotted_sectors.rs
@@ -4,7 +4,8 @@ use async_trait::async_trait;
 use futures::{stream, Stream};
 use std::sync::Arc;
 use subspace_core_primitives::pieces::PieceOffset;
-use subspace_core_primitives::{PublicKey, SectorId};
+use subspace_core_primitives::sectors::SectorId;
+use subspace_core_primitives::PublicKey;
 use subspace_farmer_components::plotting::PlottedSector;
 use subspace_farmer_components::sector::SectorMetadataChecksummed;
 use subspace_farmer_components::FarmerProtocolInfo;
