+++ shared/subspace-proof-of-space-gpu/src/cuda/tests.rs
@@ -3,8 +3,8 @@ use std::num::NonZeroUsize;
 use std::slice;
 use subspace_core_primitives::crypto::{blake3_254_hash_to_scalar, blake3_hash};
 use subspace_core_primitives::pieces::{PieceOffset, Record};
+use subspace_core_primitives::sectors::SectorId;
 use subspace_core_primitives::segments::HistorySize;
-use subspace_core_primitives::SectorId;
 use subspace_erasure_coding::ErasureCoding;
 use subspace_farmer_components::plotting::{CpuRecordsEncoder, RecordsEncoder};
 use subspace_farmer_components::sector::SectorContentsMap;
